backend/pyproject.toml
[project]
name = "invoice-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.1",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "psycopg[binary]>=3.3.2",
    "python-dotenv>=1.2.1",
]
